Feature/digital wallet os 21 (#79)
* feat: (digital wallet) backend side

* feat: (digital wallet) fronted to pay the booking

* chore: (emailing) bugfix
