+++ apps/api/src/controllers/payment/digital-payment-controller.ts
@@ -200,7 +200,7 @@ export async function createDigitalPayment(req: RequestWithUserId, res: Response
       let scheduleDate =
          now > new Date(booking.startDate)
             ? new Date().getTime() + 1000 * 10
-            : new Date(booking.startDate.getDate() - 1);
+            : new Date(booking.startDate.setDate(booking.startDate.getDate() - 1));
 
       const job = schedule.scheduleJob(scheduleDate, async () => {
          try {
@@ -236,7 +236,7 @@ export async function createDigitalPayment(req: RequestWithUserId, res: Response
          }
       });
 
-      const updateDate = new Date(new Date(booking.endDate).getDate() + 1);
+      const updateDate = new Date(booking.endDate.setDate(booking.endDate.getDate() + 1));
       const updateJob = schedule.scheduleJob(updateDate, async () => {
          try {
             await prisma.booking.update({
